refactor region view
